api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.5.0
joblib==1.4.2
shap==0.46.0
lime==0.2.0.1
apscheduler==3.10.4
httpx==0.27.2
python-multipart==0.0.12
sqlalchemy[asyncio]==2.0.36
supabase==2.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.2.2
python-dotenv==1.0.1
